requirements.txt
# Core dependencies
numpy>=1.20.0
pillow>=9.0.0
pyyaml>=6.0
requests>=2.27.0
tqdm>=4.62.0

# Vision module
opencv-python>=4.5.5.64
face-recognition>=1.3.0
fer>=22.4.0
mtcnn>=0.1.1

# Speech module
sounddevice>=0.4.5
faster-whisper>=0.9.0
llama-cpp-python>=0.2.0
piper-tts>=1.2.0

# Display module
pygame>=2.1.2

# Communication module
fastapi>=0.95.0
uvicorn>=0.21.0
websockets>=10.4

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=23.1.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
pylint>=2.16.0 